# docs/openapi.json
{
  "components": {
    "schemas": {
      "AccrualMethod": {
        "enum": [
          "TIME",
          "HOURS_WORKED"
        ],
        "title": "AccrualMethod",
        "type": "string"
      },
      "AccrualRuleIn": {
        "properties": {
          "accrues_at": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccruesAt"
              },
              {
                "type": "null"
              }
            ]
          },
          "amount": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "frequency": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Schedule"
              },
              {
                "type": "null"
              }
            ]
          },
          "method": {
            "$ref": "#/components/schemas/AccrualMethod"
          },
          "min_tenure_months": {
            "default": 0,
            "minimum": 0.0,
            "title": "Min Tenure Months",
            "type": "integer"
          },
          "per_minutes_worked": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Per Minutes Worked"
          },
          "unit": {
            "$ref": "#/components/schemas/RateUnit"
          }
        },
        "required": [
          "method",
          "amount",
          "unit"
        ],
        "title": "AccrualRuleIn",
        "type": "object"
      },
      "AccrualRuleOut": {
        "properties": {
          "accrues_at": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccruesAt"
              },
              {
                "type": "null"
              }
            ]
          },
          "amount": {
            "title": "Amount",
            "type": "string"
          },
          "frequency": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Schedule"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "method": {
            "$ref": "#/components/schemas/AccrualMethod"
          },
          "min_tenure_months": {
            "title": "Min Tenure Months",
            "type": "integer"
          },
          "per_minutes_worked": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Per Minutes Worked"
          },
          "unit": {
            "$ref": "#/components/schemas/RateUnit"
          }
        },
        "required": [
          "id",
          "method",
          "amount",
          "unit",
          "frequency",
          "accrues_at",
          "per_minutes_worked",
          "min_tenure_months"
        ],
        "title": "AccrualRuleOut",
        "type": "object"
      },
      "AccruesAt": {
        "enum": [
          "START_OF_PERIOD",
          "END_OF_PERIOD"
        ],
        "title": "AccruesAt",
        "type": "string"
      },
...
      "CategoryCreateIn": {
        "properties": {
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "name": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CategoryCreateIn",
        "type": "object"
      },
      "CategoryOut": {
        "properties": {
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "icon"
        ],
        "title": "CategoryOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NewHireProration": {
        "enum": [
          "PRORATE",
          "FULL",
          "NONE"
        ],
        "title": "NewHireProration",
        "type": "string"
      },
...
      "PolicyCreateIn": {
        "properties": {
          "all_employees": {
            "title": "All Employees",
            "type": "boolean"
          },
          "allow_negative": {
            "default": false,
            "title": "Allow Negative",
            "type": "boolean"
          },
          "carryover_cap_minutes": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carryover Cap Minutes"
          },
          "category_id": {
            "title": "Category Id",
            "type": "string"
          },
          "change_reason": {
            "default": "Policy created",
            "title": "Change Reason",
            "type": "string"
          },
          "effective_from": {
            "format": "date",
            "title": "Effective From",
            "type": "string"
          },
          "expires_at_period_end": {
            "default": false,
            "title": "Expires At Period End",
            "type": "boolean"
          },
          "group_ids": {
            "items": {
              "type": "string"
            },
            "title": "Group Ids",
            "type": "array"
          },
          "kind": {
            "$ref": "#/components/schemas/PolicyKind"
          },
          "max_balance_minutes": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Balance Minutes"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "negative_floor_minutes": {
            "default": 0,
            "maximum": 0.0,
            "title": "Negative Floor Minutes",
            "type": "integer"
          },
          "new_hire_proration": {
            "$ref": "#/components/schemas/NewHireProration",
            "default": "PRORATE"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/AccrualRuleIn"
            },
            "title": "Rules",
            "type": "array"
          },
          "tenure_transition": {
            "$ref": "#/components/schemas/TenureTransition",
            "default": "NEXT_PERIOD"
          }
        },
        "required": [
          "name",
          "category_id",
          "effective_from",
          "kind",
          "all_employees"
        ],
        "title": "PolicyCreateIn",
        "type": "object"
      },
      "PolicyKind": {
        "enum": [
          "UNLIMITED",
          "ACCRUAL"
        ],
        "title": "PolicyKind",
        "type": "string"
      },
      "PolicyOut": {
        "properties": {
          "all_employees": {
            "title": "All Employees",
            "type": "boolean"
          },
          "category_id": {
            "title": "Category Id",
            "type": "string"
          },
          "category_name": {
            "title": "Category Name",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "current_version": {
            "$ref": "#/components/schemas/PolicyVersionOut"
          },
          "group_ids": {
            "items": {
              "type": "string"
            },
            "title": "Group Ids",
            "type": "array"
          },
          "group_names": {
            "items": {
              "type": "string"
            },
            "title": "Group Names",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "version_count": {
            "title": "Version Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "category_id",
          "category_name",
          "created_by",
          "current_version",
          "version_count",
          "all_employees",
          "group_ids",
          "group_names"
        ],
        "title": "PolicyOut",
        "type": "object"
      },
      "PolicyUpdateIn": {
        "properties": {
          "all_employees": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "All Employees"
          },
          "allow_negative": {
            "default": false,
            "title": "Allow Negative",
            "type": "boolean"
          },
          "carryover_cap_minutes": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carryover Cap Minutes"
          },
          "change_reason": {
            "minLength": 1,
            "title": "Change Reason",
            "type": "string"
          },
          "effective_from": {
            "format": "date",
            "title": "Effective From",
            "type": "string"
          },
          "expires_at_period_end": {
            "default": false,
            "title": "Expires At Period End",
            "type": "boolean"
          },
          "group_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Ids"
          },
          "kind": {
            "$ref": "#/components/schemas/PolicyKind"
          },
          "max_balance_minutes": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Balance Minutes"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "negative_floor_minutes": {
            "default": 0,
            "maximum": 0.0,
            "title": "Negative Floor Minutes",
            "type": "integer"
          },
          "new_hire_proration": {
            "$ref": "#/components/schemas/NewHireProration",
            "default": "PRORATE"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/AccrualRuleIn"
            },
            "title": "Rules",
            "type": "array"
          },
          "tenure_transition": {
            "$ref": "#/components/schemas/TenureTransition",
            "default": "NEXT_PERIOD"
          }
        },
        "required": [
          "effective_from",
          "kind",
          "change_reason"
        ],
        "title": "PolicyUpdateIn",
        "type": "object"
      },
      "PolicyVersionOut": {
        "properties": {
          "allow_negative": {
            "title": "Allow Negative",
            "type": "boolean"
          },
          "carryover_cap_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carryover Cap Minutes"
          },
          "change_reason": {
            "title": "Change Reason",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "effective_from": {
            "format": "date",
            "title": "Effective From",
            "type": "string"
          },
          "expires_at_period_end": {
            "title": "Expires At Period End",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/PolicyKind"
          },
          "max_balance_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Balance Minutes"
          },
          "negative_floor_minutes": {
            "title": "Negative Floor Minutes",
            "type": "integer"
          },
          "new_hire_proration": {
            "$ref": "#/components/schemas/NewHireProration"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/AccrualRuleOut"
            },
            "title": "Rules",
            "type": "array"
          },
          "tenure_transition": {
            "$ref": "#/components/schemas/TenureTransition"
          },
          "version_no": {
            "title": "Version No",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "version_no",
          "effective_from",
          "kind",
          "created_by",
          "change_reason",
          "new_hire_proration",
          "allow_negative",
          "negative_floor_minutes",
          "max_balance_minutes",
          "carryover_cap_minutes",
          "expires_at_period_end",
          "tenure_transition",
          "created_at",
          "rules"
        ],
        "title": "PolicyVersionOut",
        "type": "object"
      },
      "RateUnit": {
        "enum": [
          "DAY",
          "HOUR",
          "MINUTE"
        ],
        "title": "RateUnit",
        "type": "string"
      },
...
      "Schedule": {
        "enum": [
          "DAILY",
          "WEEKLY",
          "SEMIMONTHLY",
          "BIWEEKLY",
          "MONTHLY",
          "YEARLY"
        ],
        "title": "Schedule",
        "type": "string"
      },
...
      "TenureTransition": {
        "enum": [
          "NEXT_PERIOD"
        ],
        "title": "TenureTransition",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Employee time off tracking: policies, accrual, requests, balances, audit.",
    "title": "Warp Time Off",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/categories": {
      "get": {
        "operationId": "categories_api_categories_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-actor-id",
            "required": false,
            "schema": {
              "default": "adm_lindsey",
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategoryOut"
                  },
                  "title": "Response Categories Api Categories Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Categories",
        "tags": [
          "policies"
        ]
      },
      "post": {
        "operationId": "create_category_api_categories_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-actor-id",
            "required": false,
            "schema": {
              "default": "adm_lindsey",
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Category",
        "tags": [
          "policies"
        ]
      }
    },
...
    "/api/policies": {
      "get": {
        "operationId": "policies_api_policies_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-actor-id",
            "required": false,
            "schema": {
              "default": "adm_lindsey",
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PolicyOut"
                  },
                  "title": "Response Policies Api Policies Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Policies",
        "tags": [
          "policies"
        ]
      },
      "post": {
        "operationId": "create_policy_api_policies_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-actor-id",
            "required": false,
            "schema": {
              "default": "adm_lindsey",
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Policy",
        "tags": [
          "policies"
        ]
      }
    },
    "/api/policies/{policy_id}": {
      "put": {
        "operationId": "update_policy_api_policies__policy_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "policy_id",
            "required": true,
            "schema": {
              "title": "Policy Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-actor-id",
            "required": false,
            "schema": {
              "default": "adm_lindsey",
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyVersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Policy",
        "tags": [
          "policies"
        ]
      }
    },
...
    "/api/policies/{policy_id}/versions": {
      "get": {
        "operationId": "versions_api_policies__policy_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "policy_id",
            "required": true,
            "schema": {
              "title": "Policy Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-actor-id",
            "required": false,
            "schema": {
              "default": "adm_lindsey",
              "title": "X-Actor-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PolicyVersionOut"
                  },
                  "title": "Response Versions Api Policies  Policy Id  Versions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Versions",
        "tags": [
          "policies"
        ]
      }
    },
...
  }
}